fix: Fix accessing project URL directly
We were setting current project based on its slug on
`ProjectContainer.mounted` method. But resources (and so, projects) are
not in store yet. To fix that, we need to set current project once
resources are loaded.

We still need to set current project on `mounted` because resources can
already be loaded.

Also, user can try to access a project that doesn't exist and so we need
to show a "not found" page in this case.